requirements.txt
asgiref==3.9.1
channels==4.3.0
colorama==0.4.6
Django==5.2.4
django-bootstrap5==25.1
django-cors-headers==4.7.0
django-debug-toolbar==6.0.0
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-js-asset==3.1.2
django-mptt==0.17.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
et_xmlfile==2.0.0
inflection==0.5.1
iniconfig==2.1.0
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
